requirements.txt
cassandra-driver
customtkinter
Pillow